+++ pulse-sys/src/mainloop/api.rs
@@ -34,45 +34,45 @@ pub mod io_event_flags {
 
 /// An opaque IO event source object
 pub enum pa_io_event {}
-pub type pa_io_event_cb_t = Option<extern "C" fn(a: *mut pa_mainloop_api, e: *mut pa_io_event, fd: i32, events: pa_io_event_flags_t, userdata: *mut c_void)>;
-pub type pa_io_event_destroy_cb_t = Option<extern "C" fn(a: *mut pa_mainloop_api, e: *mut pa_io_event, userdata: *mut c_void)>;
+pub type pa_io_event_cb_t = Option<extern "C" fn(a: *const pa_mainloop_api, e: *mut pa_io_event, fd: i32, events: pa_io_event_flags_t, userdata: *mut c_void)>;
+pub type pa_io_event_destroy_cb_t = Option<extern "C" fn(a: *const pa_mainloop_api, e: *mut pa_io_event, userdata: *mut c_void)>;
 
 /// An opaque timer event source object
 pub enum pa_time_event {}
-pub type pa_time_event_cb_t = Option<extern "C" fn(a: *mut pa_mainloop_api, e: *mut pa_time_event, tv: *const timeval, userdata: *mut c_void)>;
-pub type pa_time_event_destroy_cb_t = Option<extern "C" fn(a: *mut pa_mainloop_api, e: *mut pa_time_event, userdata: *mut c_void)>;
+pub type pa_time_event_cb_t = Option<extern "C" fn(a: *const pa_mainloop_api, e: *mut pa_time_event, tv: *const timeval, userdata: *mut c_void)>;
+pub type pa_time_event_destroy_cb_t = Option<extern "C" fn(a: *const pa_mainloop_api, e: *mut pa_time_event, userdata: *mut c_void)>;
 
 /// An opaque deferred event source object.
 /// Events of this type are triggered once in every main loop iteration
 pub enum pa_defer_event {}
-pub type pa_defer_event_cb_t = Option<extern "C" fn(a: *mut pa_mainloop_api, e: *mut pa_defer_event, userdata: *mut c_void)>;
-pub type pa_defer_event_destroy_cb_t = Option<extern "C" fn(a: *mut pa_mainloop_api, e: *mut pa_defer_event, userdata: *mut c_void)>;
+pub type pa_defer_event_cb_t = Option<extern "C" fn(a: *const pa_mainloop_api, e: *mut pa_defer_event, userdata: *mut c_void)>;
+pub type pa_defer_event_destroy_cb_t = Option<extern "C" fn(a: *const pa_mainloop_api, e: *mut pa_defer_event, userdata: *mut c_void)>;
 
 #[repr(C)]
 pub struct pa_mainloop_api {
     pub userdata: *mut c_void,
 
-    pub io_new: Option<extern "C" fn(a: *mut pa_mainloop_api, fd: i32, events: pa_io_event_flags_t, cb: pa_io_event_cb_t, userdata: *mut c_void) -> *mut pa_io_event>,
+    pub io_new: Option<extern "C" fn(a: *const pa_mainloop_api, fd: i32, events: pa_io_event_flags_t, cb: pa_io_event_cb_t, userdata: *mut c_void) -> *mut pa_io_event>,
     pub io_enable: Option<extern "C" fn(e: *mut pa_io_event, events: pa_io_event_flags_t)>,
     pub io_free: Option<extern "C" fn(e: *mut pa_io_event)>,
     pub io_set_destroy: Option<extern "C" fn(e: *mut pa_io_event, cb: pa_io_event_destroy_cb_t)>,
 
-    pub time_new: Option<extern "C" fn(a: *mut pa_mainloop_api, tv: *const timeval, cb: pa_time_event_cb_t, userdata: *mut c_void) -> *mut pa_time_event>,
+    pub time_new: Option<extern "C" fn(a: *const pa_mainloop_api, tv: *const timeval, cb: pa_time_event_cb_t, userdata: *mut c_void) -> *mut pa_time_event>,
     pub time_restart: Option<extern "C" fn(e: *mut pa_time_event, tv: *const timeval)>,
     pub time_free: Option<extern "C" fn(e: *mut pa_time_event)>,
     pub time_set_destroy: Option<extern "C" fn(e: *mut pa_time_event, cb: pa_time_event_destroy_cb_t)>,
 
-    pub defer_new: Option<extern "C" fn(a: *mut pa_mainloop_api, cb: pa_defer_event_cb_t, userdata: *mut c_void) -> *mut pa_defer_event>,
+    pub defer_new: Option<extern "C" fn(a: *const pa_mainloop_api, cb: pa_defer_event_cb_t, userdata: *mut c_void) -> *mut pa_defer_event>,
     pub defer_enable: Option<extern "C" fn(e: *mut pa_defer_event, b: i32)>,
     pub defer_free: Option<extern "C" fn(e: *mut pa_defer_event)>,
     pub defer_set_destroy: Option<extern "C" fn(e: *mut pa_defer_event, cb: pa_defer_event_destroy_cb_t)>,
 
-    pub quit: Option<extern "C" fn(a: *mut pa_mainloop_api, retval: i32)>,
+    pub quit: Option<extern "C" fn(a: *const pa_mainloop_api, retval: i32)>,
 }
 
-pub type pa_mainloop_api_once_cb = Option<extern "C" fn(m: *mut pa_mainloop_api, userdata: *mut c_void)>;
+pub type pa_mainloop_api_once_cb = Option<extern "C" fn(m: *const pa_mainloop_api, userdata: *mut c_void)>;
 
 #[link(name="pulse")]
 extern "C" {
-    pub fn pa_mainloop_api_once(m: *mut pa_mainloop_api, callback: pa_mainloop_api_once_cb, userdata: *mut c_void);
+    pub fn pa_mainloop_api_once(m: *const pa_mainloop_api, callback: pa_mainloop_api_once_cb, userdata: *mut c_void);
 }